backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic-settings
pydantic[email]
Pillow
python-magic
opencv-python
numpy
scikit-image
python-socketio
aiohttp
